+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/aad/b2c/implementation/AADB2CConditions.java
@@ -2,11 +2,8 @@
 // Licensed under the MIT License.
 package com.azure.spring.cloud.autoconfigure.aad.b2c.implementation;
 
-import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
 import org.springframework.boot.autoconfigure.condition.ConditionMessage;
 import org.springframework.boot.autoconfigure.condition.ConditionOutcome;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.autoconfigure.condition.SpringBootCondition;
 import org.springframework.boot.context.properties.bind.Binder;
 import org.springframework.context.annotation.ConditionContext;
@@ -19,39 +16,6 @@ import java.util.Map;
  * Conditions for activating AAD B2C beans.
  */
 public final class AADB2CConditions {
-
-    /**
-     * Web application or web resource server scenario condition.
-     */
-    static final class CommonCondition extends AnyNestedCondition {
-        CommonCondition() {
-            super(ConfigurationPhase.REGISTER_BEAN);
-        }
-
-        /**
-         * Web application scenario condition.
-         */
-        @ConditionalOnWebApplication
-        @ConditionalOnProperty(
-            prefix = AADB2CProperties.PREFIX,
-            value = {
-                "client-id",
-                "client-secret"
-            }
-        )
-        static class WebAppMode {
-
-        }
-
-        /**
-         * Web resource server scenario condition.
-         */
-        @ConditionalOnWebApplication
-        @ConditionalOnProperty(prefix = AADB2CProperties.PREFIX, value = { "tenant-id" })
-        static class WebApiMode {
-
-        }
-    }
 
     /**
      * OAuth2 client beans condition.