+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/aad/filter/AADAuthenticationFilter.java
@@ -113,8 +113,8 @@ public class AADAuthenticationFilter extends OncePerRequestFilter {
                                    UserPrincipalManager userPrincipalManager) {
         this.userPrincipalManager = userPrincipalManager;
         this.azureADGraphClient = new AzureADGraphClient(
-            aadAuthenticationProperties.getClientId(),
-            aadAuthenticationProperties.getClientSecret(),
+            aadAuthenticationProperties.getCredential().getClientId(),
+            aadAuthenticationProperties.getCredential().getClientSecret(),
             aadAuthenticationProperties,
             endpoints
         );