+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/aad/filter/UserPrincipalManager.java
@@ -87,7 +87,7 @@ public class UserPrincipalManager {
         this.explicitAudienceCheck = explicitAudienceCheck;
         if (explicitAudienceCheck) {
             // client-id for "normal" check
-            this.validAudiences.add(this.aadAuthenticationProperties.getClientId());
+            this.validAudiences.add(this.aadAuthenticationProperties.getCredential().getClientId());
             // app id uri for client credentials flow (server to server communication)
             this.validAudiences.add(this.aadAuthenticationProperties.getAppIdUri());
         }
@@ -123,7 +123,7 @@ public class UserPrincipalManager {
         this.explicitAudienceCheck = explicitAudienceCheck;
         if (explicitAudienceCheck) {
             // client-id for "normal" check
-            this.validAudiences.add(this.aadAuthenticationProperties.getClientId());
+            this.validAudiences.add(this.aadAuthenticationProperties.getCredential().getClientId());
             // app id uri for client credentials flow (server to server communication)
             this.validAudiences.add(this.aadAuthenticationProperties.getAppIdUri());
         }