+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/filter/ResourceRetrieverTest.java
@@ -3,6 +3,7 @@
 
 package com.azure.spring.cloud.autoconfigure.aad.filter;
 
+import com.azure.spring.cloud.autoconfigure.context.AzureGlobalPropertiesAutoConfiguration;
 import com.nimbusds.jose.jwk.source.RemoteJWKSet;
 import com.nimbusds.jose.util.DefaultResourceRetriever;
 import com.nimbusds.jose.util.ResourceRetriever;
@@ -16,12 +17,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class ResourceRetrieverTest {
     private final WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
-        .withConfiguration(AutoConfigurations.of(AADAuthenticationFilterAutoConfiguration.class))
+        .withConfiguration(AutoConfigurations.of(AzureGlobalPropertiesAutoConfiguration.class, AADAuthenticationFilterAutoConfiguration.class))
         .withClassLoader(new FilteredClassLoader(BearerTokenAuthenticationToken.class))
         .withPropertyValues(
             "spring.cloud.azure.active-directory.enabled=true",
-            "spring.cloud.azure.active-directory.client-id=fake-client-id",
-            "spring.cloud.azure.active-directory.client-secret=fake-client-secret"
+            "spring.cloud.azure.active-directory.credential.client-id=fake-client-id",
+            "spring.cloud.azure.active-directory.credential.client-secret=fake-client-secret"
         );
 
     @Test
@@ -43,7 +44,7 @@ class ResourceRetrieverTest {
     }
 
     @Test
-    void resourceRetriverIsConfigurable() {
+    void resourceRetrieverIsConfigurable() {
         this.contextRunner
             .withPropertyValues(
                 "spring.cloud.azure.active-directory.enabled=true",