+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/conditions/ResourceServerConditionTest.java
@@ -15,7 +15,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     void testResourceServerConditionWhenApplicationTypeIsEmpty() {
         this.contextRunner
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.client-id = fake-client-id")
+                "spring.cloud.azure.active-directory.credential.client-id = fake-client-id")
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(true));
     }
@@ -23,7 +23,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     @Test
     void testResourceServerConditionWhenNoOAuth2ResourceDependency() {
         this.contextRunner
-            .withPropertyValues("spring.cloud.azure.active-directory.client-id = fake-client-id")
+            .withPropertyValues("spring.cloud.azure.active-directory.credential.client-id = fake-client-id")
             .withClassLoader(new FilteredClassLoader(BearerTokenAuthenticationToken.class))
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(false));
@@ -33,7 +33,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     void testResourceServerConditionWhenApplicationTypeIsWebApplication() {
         this.contextRunner
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.client-id = fake-client-id",
+                "spring.cloud.azure.active-directory.credential.client-id = fake-client-id",
                 "spring.cloud.azure.active-directory.application-type=web_application")
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(false));
@@ -43,7 +43,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     void testResourceServerConditionWhenApplicationTypeIsResourceServer() {
         this.contextRunner
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.client-id = fake-client-id",
+                "spring.cloud.azure.active-directory.credential.client-id = fake-client-id",
                 "spring.cloud.azure.active-directory.application-type=resource_server")
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(true));
@@ -53,7 +53,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     void testResourceServerConditionWhenApplicationTypeIsResourceServerWithOBO() {
         this.contextRunner
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.client-id = fake-client-id",
+                "spring.cloud.azure.active-directory.credential.client-id = fake-client-id",
                 "spring.cloud.azure.active-directory.application-type=resource_server_with_obo")
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(true));
@@ -63,7 +63,7 @@ class ResourceServerConditionTest extends AbstractCondition {
     void testResourceServerConditionWhenApplicationTypeIsWebApplicationAndResourceServer() {
         this.contextRunner
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.client-id = fake-client-id",
+                "spring.cloud.azure.active-directory.credential.client-id = fake-client-id",
                 "spring.cloud.azure.active-directory.application-type=web_application_and_resource_server")
             .withUserConfiguration(ResourceServerConditionConfig.class)
             .run(assertConditionMatch(true));