+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/properties/AADAuthenticationPropertiesTest.java
@@ -6,6 +6,7 @@ package com.azure.spring.cloud.autoconfigure.aad.properties;
 import com.azure.spring.cloud.autoconfigure.aad.core.AADApplicationType;
 import org.junit.jupiter.api.Test;
 
+import static com.azure.spring.cloud.autoconfigure.aad.implementation.WebApplicationContextRunnerUtils.oauthClientRunner;
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.WebApplicationContextRunnerUtils.resourceServerContextRunner;
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.WebApplicationContextRunnerUtils.resourceServerWithOboContextRunner;
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.WebApplicationContextRunnerUtils.webApplicationContextRunner;
@@ -28,71 +29,56 @@ class AADAuthenticationPropertiesTest {
         webApplicationContextRunner()
             .run(context -> {
                 AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
-                assertEquals(properties.getGraphBaseUri(), "https://graph.microsoft.com/");
+                assertEquals(properties.getProfile().getEnvironment().getMicrosoftGraphEndpoint(), "https://graph.microsoft.com/");
                 assertEquals(properties.getGraphMembershipUri(), "https://graph.microsoft.com/v1.0/me/memberOf");
             });
 
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.graph-base-uri=https://microsoftgraph.chinacloudapi.cn"
+                "spring.cloud.azure.active-directory.profile.environment.microsoft-graph-endpoint=https://microsoftgraph.chinacloudapi.cn"
             )
             .run(context -> {
                 AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
-                assertEquals(properties.getGraphBaseUri(), "https://microsoftgraph.chinacloudapi.cn/");
+                assertEquals(properties.getProfile().getEnvironment().getMicrosoftGraphEndpoint(), "https://microsoftgraph.chinacloudapi.cn/");
                 assertEquals(properties.getGraphMembershipUri(),
                     "https://microsoftgraph.chinacloudapi.cn/v1.0/me/memberOf");
             });
 
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.graph-base-uri=https://microsoftgraph.chinacloudapi.cn/"
+                "spring.cloud.azure.active-directory..profile.environment.microsoft-graph-endpoint=https://microsoftgraph.chinacloudapi.cn/"
             )
             .run(context -> {
                 AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
-                assertEquals(properties.getGraphBaseUri(), "https://microsoftgraph.chinacloudapi.cn/");
+                assertEquals(properties.getProfile().getEnvironment().getMicrosoftGraphEndpoint(), "https://microsoftgraph.chinacloudapi.cn/");
                 assertEquals(properties.getGraphMembershipUri(),
                     "https://microsoftgraph.chinacloudapi.cn/v1.0/me/memberOf");
             });
 
         webApplicationContextRunner()
-            .withPropertyValues(
-                "spring.cloud.azure.active-directory.graph-membership-uri=https://graph.microsoft.com/v1.0/me/memberOf"
-            )
             .run(context -> {
                 AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
-                assertEquals(properties.getGraphBaseUri(), "https://graph.microsoft.com/");
+                assertEquals(properties.getProfile().getEnvironment().getMicrosoftGraphEndpoint(), "https://graph.microsoft.com/");
                 assertEquals(properties.getGraphMembershipUri(), "https://graph.microsoft.com/v1.0/me/memberOf");
             });
 
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.graph-base-uri=https://microsoftgraph.chinacloudapi.cn/",
-                "spring.cloud.azure.active-directory.graph-membership-uri=https://microsoftgraph.chinacloudapi.cn/v1.0/me/memberOf"
+                "spring.cloud.azure.active-directory.profile.environment.microsoft-graph-endpoint=https://microsoftgraph.chinacloudapi.cn/"
             )
             .run(context -> {
                 AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
-                assertEquals(properties.getGraphBaseUri(), "https://microsoftgraph.chinacloudapi.cn/");
+                assertEquals(properties.getProfile().getEnvironment().getMicrosoftGraphEndpoint(), "https://microsoftgraph.chinacloudapi.cn/");
                 assertEquals(properties.getGraphMembershipUri(),
                     "https://microsoftgraph.chinacloudapi.cn/v1.0/me/memberOf");
             });
-    }
-
-    @Test
-    void graphUriConfigurationWithExceptionTest() {
-        webApplicationContextRunner()
-            .withPropertyValues(
-                "spring.cloud.azure.active-directory.graph-membership-uri=https://microsoftgraph.chinacloudapi.cn/v1.0/me/memberOf"
-            )
-            .run(context ->
-                assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class))
-            );
     }
 
     @Test
     void multiTenantWithAllowedGroupsConfiguredTest1() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=",
+                "spring.cloud.azure.active-directory.profile.tenant-id=",
                 "spring.cloud.azure.active-directory.user-group.allowed-groups=group1,group2"
             )
             .run(context ->
@@ -104,7 +90,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsConfiguredTest2() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=common",
+                "spring.cloud.azure.active-directory.profile.tenant-id=common",
                 "spring.cloud.azure.active-directory.user-group.allowed-groups=group1,group2"
             )
             .run(context ->
@@ -116,7 +102,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsConfiguredTest3() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=organizations",
+                "spring.cloud.azure.active-directory.profile.tenant-id=organizations",
                 "spring.cloud.azure.active-directory.user-group.allowed-groups=group1,group2"
             )
             .run(context ->
@@ -128,7 +114,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsIdConfiguredTest1() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=",
+                "spring.cloud.azure.active-directory.profile.tenant-id=",
                 "spring.cloud.azure.active-directory.user-group.allowed-group-ids = 7c3a5d22-9093-42d7-b2eb-e72d06bf3718,"
                     + "39087533-2593-4b5b-ad05-4a73a01ea6a9"
             )
@@ -141,7 +127,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsIdConfiguredTest2() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=common",
+                "spring.cloud.azure.active-directory.profile.tenant-id=common",
                 "spring.cloud.azure.active-directory.user-group.allowed-group-ids = 7c3a5d22-9093-42d7-b2eb-e72d06bf3718,"
                     + "39087533-2593-4b5b-ad05-4a73a01ea6a9"
             )
@@ -154,7 +140,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsIdConfiguredTest3() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=organizations",
+                "spring.cloud.azure.active-directory.profile.tenant-id=organizations",
                 "spring.cloud.azure.active-directory.user-group.allowed-group-ids = 7c3a5d22-9093-42d7-b2eb-e72d06bf3718,"
                     + "39087533-2593-4b5b-ad05-4a73a01ea6a9"
             )
@@ -167,7 +153,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsIdConfiguredTest4() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=consumers",
+                "spring.cloud.azure.active-directory.profile.tenant-id=consumers",
                 "spring.cloud.azure.active-directory.user-group.allowed-group-ids = 7c3a5d22-9093-42d7-b2eb-e72d06bf3718,"
                     + "39087533-2593-4b5b-ad05-4a73a01ea6a9"
             )
@@ -180,7 +166,7 @@ class AADAuthenticationPropertiesTest {
     void multiTenantWithAllowedGroupsConfiguredTest4() {
         webApplicationContextRunner()
             .withPropertyValues(
-                "spring.cloud.azure.active-directory.tenant-id=consumers",
+                "spring.cloud.azure.active-directory.profile.tenant-id=consumers",
                 "spring.cloud.azure.active-directory.user-group.allowed-groups=group1,group2"
             )
             .run(context ->
@@ -275,15 +261,11 @@ class AADAuthenticationPropertiesTest {
                 "spring.cloud.azure.active-directory.enabled=true",
                 "spring.cloud.azure.active-directory.application-type=web_application"
             )
-            .run(context -> {
-                assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class));
-            });
+            .run(context -> assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class)));
 
         webApplicationContextRunner()
             .withPropertyValues("spring.cloud.azure.active-directory.application-type=resource_server")
-            .run(context -> {
-                assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class));
-            });
+            .run(context -> assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class)));
     }
 
     @Test
@@ -294,8 +276,39 @@ class AADAuthenticationPropertiesTest {
                 "spring.cloud.azure.active-directory.authorization-clients.graph.on-demand = true",
                 "spring.cloud.azure.active-directory.authorization-clients.graph.authorizationGrantType = azure_delegated"
             )
+            .run(context -> assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class)));
+    }
+
+    @Test
+    void setDefaultValueFromAzureGlobalPropertiesTest() {
+        oauthClientRunner()
+            .withPropertyValues(
+                "spring.cloud.azure.active-directory.enabled = true",
+                "spring.cloud.azure.credential.client-id = global-client-id",
+                "spring.cloud.azure.credential.client-secret = global-client-secret",
+                "spring.cloud.azure.profile.tenant-id = global-tenant-id",
+                "spring.cloud.azure.active-directory.credential.client-id = aad-client-id",
+                "spring.cloud.azure.active-directory.credential.client-secret = aad-client-secret",
+                "spring.cloud.azure.active-directory.profile.tenant-id = aad-tenant-id"
+            )
             .run(context -> {
-                assertThrows(IllegalStateException.class, () -> context.getBean(AADAuthenticationProperties.class));
+                AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
+                assertEquals("aad-client-id", properties.getCredential().getClientId());
+                assertEquals("aad-client-secret", properties.getCredential().getClientSecret());
+                assertEquals("aad-tenant-id", properties.getProfile().getTenantId());
+            });
+        oauthClientRunner()
+            .withPropertyValues(
+                "spring.cloud.azure.active-directory.enabled = true",
+                "spring.cloud.azure.credential.client-id = global-client-id",
+                "spring.cloud.azure.credential.client-secret = global-client-secret",
+                "spring.cloud.azure.profile.tenant-id = global-tenant-id"
+            )
+            .run(context -> {
+                AADAuthenticationProperties properties = context.getBean(AADAuthenticationProperties.class);
+                assertEquals("global-client-id", properties.getCredential().getClientId());
+                assertEquals("global-client-secret", properties.getCredential().getClientSecret());
+                assertEquals("global-tenant-id", properties.getProfile().getTenantId());
             });
     }
 }